add board
This is a simple board for the Sparkfun Thing to provide power and
ground rails and two I2C buses.  The board was designed with Eagle,
but the directory includes the gerber files in buses.zip, so you do
not need to run proprietary software to have the board manufactured.

I will redo it with KiCAD when I have some time.